Use different basenames for variants of link tests
Use different names for files and links created by the
variants of the link test, so that any damaged directories
created can be tracked to the test that created them.

Fix tabs in that area of the code to reflect 4 spaces per indent

Record the return code of unlink so it may be reported in
debugging code if desired.

Prep for tagging version 1.17